feat(datasee): fix comments, move dataseedInitializer
